gain percent to gain

+++ src/changes/GainChange.hpp
@@ -6,13 +6,13 @@ class SongEditor;
 
 class GainChange : public QUndoCommand {
   SongEditor *const song_editor_pointer;
-  const int old_value;
-  int new_value;
+  const double old_value;
+  double new_value;
 
 public:
   explicit GainChange(SongEditor *song_editor_pointer_input,
-                        int old_value_input,
-                        int new_value_input);
+                        double old_value_input,
+                        double new_value_input);
 
   [[nodiscard]] auto id() const -> int override;
   [[nodiscard]] auto
